dynamic select
I continue to work in the same repo as yesterday and add a select to filter the results based on the archetype.
When the user selects a value from the list, a call is made to the API with the selected archetype

Bonuses:
Create a component that shows the total number of results obtained.